Start transition to representing all objects as ConsCell ptrs.
Start the move away from distinct C++ classes to represent Sexp, Cons,
Symbol, etc. Everything is now just a pointer to a ConsCell.

This uglies up the code quite a bit and means we lose the type safety
and other advantages of real C++ classes in order to be more faithful
to the description of the implementation from McCarthy's paper.