Move Intern and Init into the reader:: namespace.
Since the symbol table is no longer owned by the reader object, it
doesn't make sense to have Intern be a member function. Also, move
Init() into the module scope and handle it the same as the Init function
for all other modules. This also removes the need to call mclisp::Init
before instantiating the reader object.